GOVSI-1114: Implement specific TokenSigning extension
- Create a new extension `TokenSigningExtension` that extends `KmsKeyExtension`
- Move specific token signing and verification code to the new extension
- Leave the base `KmsKeyExtension` generic as this will also be used to provide audit signing keys
